+++ dev-kubernetes-manifests/balance-reader.yaml
@@ -41,7 +41,7 @@ spec:
           value: "8080"
         # toggle Cloud Trace export
         - name: ENABLE_TRACING
-          value: "false"
+          value: "true"
         - name: POLL_MS
           value: "100"
         - name: CACHE_SIZE
