+++ dev-kubernetes-manifests/transaction-history.yaml
@@ -40,7 +40,7 @@ spec:
         - name: PORT
           value: "8080"
         - name: ENABLE_TRACING
-          value: "false"
+          value: "true"
         - name: POLL_MS
           value: "100"
         - name: CACHE_SIZE
