+++ src/balancereader/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.port = ${PORT}
 handlers = java.util.logging.ConsoleHandler
 spring.main.banner-mode=off
-# tracing on/off toggle 
+# tracing on/off toggle
 spring.sleuth.enabled = ${ENABLE_TRACING}
 # if tracing enabled, configuration options:
 spring.sleuth.sampler.probability=1.0
